List on the screen, but not filtering yet.
Squashing commits to allow easier rebasing to master.

* initial reorganization
* compiling, base functionality still present
* tests compiling
* Create new CredentialItemListView and Presenter
* Move createItemListDriver into base class
* Add single credential fill
* Temporarily move filtering to get the item list on the screen.
* Prepare insertion of filtering identifiers, and extraction of selected login.
* Tidy.
