+++ Shared/Common/Extensions/BaseUserDefaults+.swift
@@ -17,4 +17,9 @@ extension UserDefaults {
         return self.on(setting: UserDefaultKey.autoLockTime.rawValue, type: String.self)
                 .map { Setting.AutoLock(rawValue: $0) ?? Constant.setting.defaultAutoLock }
     }
+
+    var onItemListSort: Observable<Setting.ItemListSort> {
+        return self.on(setting: UserDefaultKey.itemListSort.rawValue, type: String.self)
+            .map { Setting.ItemListSort(rawValue: $0) ?? Constant.setting.defaultItemListSort }
+    }
 }
