+++ Shared/Store/BaseUserDefaultStore.swift
@@ -15,6 +15,10 @@ class BaseUserDefaultStore {
         return self.userDefaults.onAutoLockTime
     }
 
+    public var itemListSort: Observable<Setting.ItemListSort> {
+        return self.userDefaults.onItemListSort
+    }
+
     init(dispatcher: Dispatcher = Dispatcher.shared,
          sharedUserDefaults: UserDefaults = UserDefaults(suiteName: Constant.app.group) ?? UserDefaults.standard) {
         self.dispatcher = dispatcher
