+++ lockbox-ios/Common/Resources/Constants.swift
@@ -29,7 +29,6 @@ extension Constant.fxa {
 extension Constant.setting {
     static let defaultPreferredBrowser = Setting.PreferredBrowser.Safari
     static let defaultRecordUsageData = true
-    static let defaultItemListSort = Setting.ItemListSort.alphabetically
 }
 
 extension Constant.string {
@@ -37,7 +36,6 @@ extension Constant.string {
     static let alphabetically = NSLocalizedString("alphabetically", value: "Alphabetically", comment: "Label for the option sheet action allowing users to sort an entry list alphabetically")
     static let aToZ = NSLocalizedString("a_to_z", value: "A-Z", comment: "Label for the button allowing users to sort an entry list alphabetically")
     static let back = NSLocalizedString("back", value: "Back", comment: "Back button title")
-    static let cancel = NSLocalizedString("cancel", value: "Cancel", comment: "Cancel button title")
     static let close = NSLocalizedString("close", value: "Close", comment: "Close button title")
     static let unlink = NSLocalizedString("unlink", value: "Disconnect", comment: "Unlink aka Disconnect button title")
     static let done = NSLocalizedString("done", value: "Done", comment: "Text on button to close settings")
@@ -54,9 +52,7 @@ extension Constant.string {
     static let sortEntries = NSLocalizedString("sort_entries", value: "Sort Entries", comment: "Title for the option sheet allowing users to sort entries")
     static let unnamedEntry = NSLocalizedString("unnamed_entry", value: "unnamed entry", comment: "Placeholder text for when there is no entry name")
     static let username = NSLocalizedString("username", value: "Username", comment: "Section title text for username on the item detail screen")
-    static let usernamePlaceholder = NSLocalizedString("username_placeholder", value: "(no username)", comment: "Placeholder text when there is no username")
     static let webAddress = NSLocalizedString("web_address", value: "Web Address", comment: "Section title text for the web address on the item detail screen")
-    static let yourLockbox = NSLocalizedString("your_lockbox", value: "Your Firefox Lockbox", comment: "Title appearing above the list of entries on the main screen of the app")
     static let settingsSupportSectionHeader = NSLocalizedString("settings.support.header", value: "SUPPORT", comment: "Support section label in settings")
     static let settingsConfigurationSectionHeader = NSLocalizedString("settings.configuration.header", value: "CONFIGURATION", comment: "Configuration label in settings")
     static let settingsTitle = NSLocalizedString("settings.title", value: "Settings", comment: "Title on settings screen")
